+++ client/src/client/mod.rs
@@ -31,9 +31,15 @@ pub use self::client::{Client, SyncClient};
 pub use self::client::SecureSyncClient;
 pub use self::client_connection::ClientConnection;
 #[allow(deprecated)]
-pub use self::client_future::{ClientFuture, BasicClientHandle, ClientHandle, StreamHandle,
-                              ClientStreamHandle};
+pub use self::client_future::{ClientFuture, BasicClientHandle, ClientHandle};
 pub use self::memoize_client_handle::MemoizeClientHandle;
 pub use self::retry_client_handle::RetryClientHandle;
 #[cfg(any(feature = "openssl", feature = "ring"))]
 pub use self::secure_client_handle::SecureClientHandle;
+
+/// This is an alias for [`trust_dns_proto::StreamHandle`]
+#[deprecated(note = "use [`trust_dns_proto::StreamHandle`] instead")]
+pub use trust_dns_proto::StreamHandle;
+/// This is an alias for [`trust_dns_proto::DnsStreamHandle`]
+#[deprecated(note = "use [`trust_dns_proto::DnsStreamHandle`] instead")]
+pub use trust_dns_proto::DnsStreamHandle as ClientStreamHandle;